+++ flmodules/src/random_connections/mod.rs
@@ -1,3 +1,4 @@
+pub mod broker;
 pub mod messages;
 pub mod nodes;
 pub mod storage;
