--- flnode/src/modules/mod.rs
@@ -1,2 +1,0 @@
-pub mod random;
-pub mod stat;
